reorg(transforms) moved flatten.js up to src/utils/flatten.js
